Add `SettingsFlow` widget; Tidy up
Move settings page to settings folder
Move Your profile page to own file
Remove login button in Stats page
