+++ lib/widgets/pages/home_page.dart
@@ -5,10 +5,10 @@ import 'package:coffeecard/cubits/receipt/receipt_cubit.dart';
 import 'package:coffeecard/data/repositories/receipt_repository.dart';
 import 'package:coffeecard/service_locator.dart';
 import 'package:coffeecard/widgets/components/helpers/lazy_indexed_stack.dart';
-import 'package:coffeecard/widgets/pages/settings_page.dart';
 import 'package:coffeecard/widgets/pages/stats_page.dart';
 import 'package:coffeecard/widgets/pages/tickets_page.dart';
 import 'package:coffeecard/widgets/routers/receipts_flow.dart';
+import 'package:coffeecard/widgets/routers/settings_flow.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -39,7 +39,7 @@ class _HomePageState extends State<HomePage> {
             const TicketsPage(),
             ReceiptsFlow(),
             StatsPage(),
-            SettingsPage(),
+            SettingsFlow(),
           ],
         ),
         bottomNavigationBar: BottomNavigationBar(
